NIFI-5537 Renamed bundle based on review comments
