+++ nifi-assembly/pom.xml
@@ -399,7 +399,7 @@ language governing permissions and limitations under the License. -->
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-neo4jcypher-nar</artifactId>
+            <artifactId>nifi-neo4j-nar</artifactId>
             <version>1.8.0-SNAPSHOT</version>
             <type>nar</type>
         </dependency>
