+++ nifi-nar-bundles/nifi-neo4j-bundle/nifi-neo4j-nar/src/main/resources/META-INF/NOTICE
@@ -1,0 +1,31 @@
+nifi-neo4j-nar
+Copyright 2016 The Apache Software Foundation
+
+This product includes software developed at
+The Apache Software Foundation (http://www.apache.org/).
+
+===========================================
+Apache Software License v2
+===========================================
+
+The following binary components are provided under the Apache Software License v2
+
+  (ASL2 License) 
+    (ASLv2) Neo4j Java Driver
+      Neo4j
+      Copyright © 2002-2018 Neo4j Sweden AB (referred to in this notice as "Neo4j")
+      [http://neo4j.com]
+
+  (ASLv2) Apache Commons IO
+    The following NOTICE information applies:
+      Apache Commons IO
+      Copyright 2002-2016 The Apache Software Foundation
+
+  (ASLv2) Apache Commons Lang
+    The following NOTICE information applies:
+      Apache Commons Lang
+      Copyright 2001-2015 The Apache Software Foundation
+
+  (ASLv2) Google GSON
+    The following NOTICE information applies:
+    Copyright 2008 Google Inc.
