+++ nifi-nar-bundles/pom.xml
@@ -75,7 +75,7 @@
         <module>nifi-ignite-bundle</module>
         <module>nifi-rethinkdb-bundle</module>
         <module>nifi-influxdb-bundle</module>
-        <module>nifi-neo4jcypher-bundle</module>
+        <module>nifi-neo4j-bundle</module>
         <module>nifi-email-bundle</module>
         <module>nifi-groovyx-bundle</module>
     	<module>nifi-ranger-bundle</module>
